mtd: be silent when mtd partition parser cannot be found
Currently when we register partitions in 'parse_mtd_partitions()' we accept the
list of parsers we should try. And if one of the parsers was not found we print
a message. Well, first of all this whole idea is bad - look at how many
'part_probes' and 'part_probe_types' variables we have - nearly every driver
defines one. Instead, we should just go through all registered parsers all the
time. But this needs to be worked on separately.

This patch makes life of MTD partitions' users a bit simpler and allows them to
safely request parsers which have not been registered -
'parse_mtd_partitions()' will not print a "not available" message in this
case.

The point is that drivers do not have to do things like this any longer:

static const char *part_probe_types[] = { "cmdlinepart", "RedBoot",
                                         "afs",
                                         NULL };

but can simply do like this:

static const char *part_probe_types[] = { "cmdlinepart", "RedBoot", "afs", NULL };

Signed-off-by: Artem Bityutskiy <Artem.Bityutskiy@nokia.com>
Signed-off-by: David Woodhouse <David.Woodhouse@intel.com>